+++ tests/cmake/AddTestMaps.cmake
@@ -178,6 +178,7 @@ list(APPEND LUA_TEST_MAPS
   "bugs/1471_pickable_wrong_layer"
   "bugs/1476_crash_command_binding"
   "bugs/1494_no_error_nonexistant_model"
+  "bugs/1503_crash_invalid_save_file"
   "bugs/1508_carried_object_height"
   "bugs/1524_push_delay"
   "bugs/1526_destructible_cut_method"